Struct iecclib::Curve

#[repr(C)]

pub struct Curve {}

A curve

Methods

impl Curve

fn list() -> &'static [&'static str]

List all available hash function names.

fn lookup(name: &str) -> Result<&'static Curve, ()>

Return object representing the curve name.

This method fails if the curve named is not supported.

fn name(&self) -> &'static str

Get name of the curve.

fn field(&self) -> &'static Field

Get the field underlying the curve.

fn storage_octets(&self) -> usize

Get number of octets storing curve point takes.

fn order(&self) -> &'static [u8]

Get the order of the curve as little-endian integer.

fn clog2l(&self) -> usize

Get ceil(log2(x)) of the order of the curve.

fn rlog2l(&self) -> usize

Get round(log2(x)) of the order of the curve.

fn partial_max_bits(&self) -> usize

Get The maximum number of bits in scalar that reduce_partial can reduce.

fn scalar_elems_single(&self) -> usize

Get number of elements in single-width scalar.

fn scalar_elems_double(&self) -> usize

Get number of elements in double-width scalar (always 2 times the number in single-width).

fn scalar_octets(&self) -> usize

Get number of octets serializing a scalar takes.

fn cofactor(&self) -> u32

Get the cofactor of the curve.

fn flags(&self) -> u32

Get the flags for the curve.

fn montgomery_base(&self) -> &'static [u8]

Get the serialization of standard montgomery basepoint.

fn reduce_partial(&self, output: &mut [u32], input: &[u32]) -> Result<(), ()>

Partially reduce a scalar.

This reduces a scalar from double width to single width (possibly not fully).

The reduced value is small enough to add with single width, followed by multiply to double width.

Fails if source or target has the wrong number of elements.

fn reduce_full(&self, output: &mut [u8], input: &[u32]) -> Result<(), ()>

Fully reduce and serialize a scalar.

This fully reduces a single-width scalar and then serializes it.

Fails if source is not single-width scalar or if target is not scalar storage octets elements.

fn make_zero(&'static self) -> Point

Create a zero point.

fn make_base(&'static self) -> Point

Create a curve standard basepoint.

fn make(&'static self, val: &[u8]) -> Result<Point, ()>

Create a curve point with specified serialized value.

Fails if the point is invalid, or wrong number of octets.

fn montgomery_mul(&self,
output: &mut FieldElement,
mul: &[u32],
base: &FieldElement)
-> Result<(), ()>

Perform Montgomery multiplication.

Fails if mul is not single-width scalar or if the curve does not support montgomery form.

fn xcurve_base(&self, pubkey: &mut [u8], secret: &[u8]) -> Result<(), ()>

Perform X* DHF with base point.

Fails if pubkey or secret is not field serialization sized.

fn xcurve(&self,
shared: &mut [u8],
secret: &[u8],
ppubkey: &[u8])
-> Result<(), ()>

Perform X* DHF with peer key.

Fails if shared, secret or ppubkey is not field serialization sized, or if result of operation is all zeroes (bad peer key).