Struct iecclib::Field

#[repr(C)]

pub struct Field {}

A field

Methods

impl Field

fn list() -> &'static [&'static str]

List all available hash function names.

fn lookup(name: &str) -> Result<&'static Field, ()>

Return object representing the hash function name.

This method fails if the hash named is not supported.

fn name(&self) -> &'static str

Get name of the field.

fn field_bits(&self) -> usize

Get number of bits in field.

fn storage_octets(&self) -> usize

Get the storage size of field element in octets.

fn field_element_count(&self) -> &'static [u8]

Get the number of elements in the field.

fn field_characteristic(&self) -> &'static [u8]

Get the characteristic of the field.

fn degree(&self) -> usize

Get the degree of the field.

fn degree_stride(&self) -> usize

Get stride between the degrees of the field element.

fn degree_bits(&self) -> usize

Get number of bits in each degree.

fn check(&self, val: &[u8]) -> Result<(), ()>

Check if field element is in range.

Fails if val is wrong number of octets or out of range.

fn make_zero(&'static self) -> FieldElement

Make a zero element of this field.

fn make_int(&'static self, val: u32) -> FieldElement

Make an element with given small value (in zeroth degree) of this field.

fn make(&'static self, val: &[u8]) -> Result<FieldElement, ()>

Make an element with given serialized value.